Revert "vmm: make landlock configs VMM-level config"
This reverts commit 94929889ac2489015d60ca1480a5d412e2792c57.
This revert moves landlock config back to VMConfig.
