+++ fuzz/fuzz_targets/http_api.rs
@@ -190,6 +190,8 @@ impl RequestHandler for StubApiRequestHandler {
                 platform: None,
                 tpm: None,
                 preserved_fds: None,
+                landlock_enable: false,
+                landlock_config: None,
             })),
             state: VmState::Running,
             memory_actual_size: 0,
